#117 - reducing StationLandingPadSize intermodule communications class amount
