+++ trade-module/src/test/java/io/edpn/backend/trade/application/service/CleanUpObsoleteStationLandingPadSizeRequestsUseCaseTest.java
@@ -1,19 +1,30 @@
 package io.edpn.backend.trade.application.service;
 
+import com.fasterxml.jackson.databind.ObjectMapper;
 import io.edpn.backend.messageprocessorlib.application.dto.eddn.data.StationDataRequest;
 import io.edpn.backend.trade.application.domain.Station;
 import io.edpn.backend.trade.application.domain.System;
 import io.edpn.backend.trade.application.domain.filter.FindStationFilter;
+import io.edpn.backend.trade.application.dto.web.object.mapper.MessageMapper;
+import io.edpn.backend.trade.application.port.outgoing.kafka.SendKafkaMessagePort;
+import io.edpn.backend.trade.application.port.outgoing.station.LoadOrCreateBySystemAndStationNamePort;
 import io.edpn.backend.trade.application.port.outgoing.station.LoadStationsByFilterPort;
+import io.edpn.backend.trade.application.port.outgoing.station.UpdateStationPort;
+import io.edpn.backend.trade.application.port.outgoing.stationlandingpadsizerequest.CleanUpObsoleteStationLandingPadSizeRequestsUseCase;
+import io.edpn.backend.trade.application.port.outgoing.stationlandingpadsizerequest.CreateStationLandingPadSizeRequestPort;
 import io.edpn.backend.trade.application.port.outgoing.stationlandingpadsizerequest.DeleteStationLandingPadSizeRequestPort;
+import io.edpn.backend.trade.application.port.outgoing.stationlandingpadsizerequest.ExistsStationLandingPadSizeRequestPort;
 import io.edpn.backend.trade.application.port.outgoing.stationlandingpadsizerequest.LoadAllStationLandingPadSizeRequestsPort;
+import io.edpn.backend.trade.application.port.outgoing.system.LoadOrCreateSystemByNamePort;
 import java.util.Collections;
 import java.util.List;
+import java.util.concurrent.Executor;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
+import org.springframework.retry.support.RetryTemplate;
 
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.equalTo;
@@ -28,18 +39,50 @@ public class CleanUpObsoleteStationLandingPadSizeRequestsUseCaseTest {
 
     @Mock
     private LoadStationsByFilterPort loadStationsByFilterPort;
-
     @Mock
     private LoadAllStationLandingPadSizeRequestsPort loadAllStationLandingPadSizeRequestsPort;
-
+    @Mock
+    private LoadOrCreateSystemByNamePort loadOrCreateSystemByNamePort;
+    @Mock
+    private LoadOrCreateBySystemAndStationNamePort loadOrCreateBySystemAndStationNamePort;
+    @Mock
+    private ExistsStationLandingPadSizeRequestPort existsStationLandingPadSizeRequestPort;
+    @Mock
+    private CreateStationLandingPadSizeRequestPort createStationLandingPadSizeRequestPort;
     @Mock
     private DeleteStationLandingPadSizeRequestPort deleteStationLandingPadSizeRequestPort;
+    @Mock
+    private UpdateStationPort updateStationPort;
+    @Mock
+    private SendKafkaMessagePort sendKafkaMessagePort;
+    @Mock
+    private RetryTemplate retryTemplate;
+    @Mock
+    private Executor executor;
+    @Mock
+    private ObjectMapper objectMapper;
+    @Mock
+    private MessageMapper messageMapper;
 
-    private CleanUpObsoleteStationLandingPadSizeRequestsService underTest;
+    private CleanUpObsoleteStationLandingPadSizeRequestsUseCase underTest;
 
     @BeforeEach
     public void setUp() {
-        underTest = new CleanUpObsoleteStationLandingPadSizeRequestsService(loadStationsByFilterPort, loadAllStationLandingPadSizeRequestsPort, deleteStationLandingPadSizeRequestPort);
+        underTest = new StationLandingPadSizeInterModuleCommunicationService(
+                loadStationsByFilterPort,
+                loadAllStationLandingPadSizeRequestsPort,
+                loadOrCreateSystemByNamePort,
+                loadOrCreateBySystemAndStationNamePort,
+                existsStationLandingPadSizeRequestPort,
+                createStationLandingPadSizeRequestPort,
+                deleteStationLandingPadSizeRequestPort,
+                updateStationPort,
+                sendKafkaMessagePort,
+                retryTemplate,
+                executor,
+                objectMapper,
+                messageMapper
+        );
     }
 
     @Test
@@ -48,7 +91,7 @@ public class CleanUpObsoleteStationLandingPadSizeRequestsUseCaseTest {
                 .hasLandingPadSize(false)
                 .build();
 
-        assertThat(CleanUpObsoleteStationLandingPadSizeRequestsService.FIND_STATION_FILTER, equalTo(findStationFilter));
+        assertThat(StationLandingPadSizeInterModuleCommunicationService.FIND_STATION_FILTER, equalTo(findStationFilter));
     }
 
     @Test
